% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize.R
\name{discretize}
\alias{discretize}
\title{Spatial discretization.}
\usage{
discretize(x, n, method)
}
\arguments{
\item{x}{A numeric vector to be discretized}

\item{n}{A number of breaks}

\item{method}{A character of discretization method}
}
\value{
A vector of discretized variable of \code{x}.
}
\description{
Spatial discretization.
}
\examples{
x.disc <- discretize(x = runif(12), n = 3, method = "quantile")
table(x.disc)

}
