% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loessoptidisc.R
\name{loessoptidisc}
\alias{loessoptidisc}
\title{Strategy 2: Optimal spatial data discretization for individual variables
based on SPADE model.}
\usage{
loessoptidisc(x, y)
}
\arguments{
\item{x}{A numeric vector of break numbers}

\item{y}{A numeric vector of q values}
}
\value{
A list of an optimal number of discretization and a plot.
}
\description{
Strategy 2: Optimal spatial data discretization for individual variables
based on SPADE model.
}
\examples{
lod <- loessoptidisc(x = 4:15, y = log(4:15 + runif(12)))

}
