% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optidiscqs1.R
\name{optidiscqs1}
\alias{optidiscqs1}
\title{Strategy 1: Optimal spatial data discretization for individual variables
based on SPADE model.}
\usage{
optidiscqs1(y, x, location, ndisc, methoddisc)
}
\arguments{
\item{y}{A numeric vector of a response variable}

\item{x}{A numeric vector of a explanatory variable}

\item{location}{A matrix of spatial locations}

\item{ndisc}{A number of discretization}

\item{methoddisc}{A character of discretization methods}
}
\value{
A list of an optimal spatial discretization using strategy 1.
}
\description{
Strategy 1: Optimal spatial data discretization for individual variables
based on SPADE model.
}
\examples{
od <- optidiscqs1(y = sim[, 1], x = sim[, 4:6], location = sim[, 2:3],
                  ndisc = c(3:5), methoddisc = c("quantile", "equal"))

}
