% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectgd.R
\name{selectgd}
\alias{selectgd}
\title{Selecting optimal interaction for GD model.}
\usage{
selectgd(formula, data, ndisc, methoddisc)
}
\arguments{
\item{formula}{A formula of spatial variables}

\item{data}{A data frame of dataset}

\item{ndisc}{A numeric vector of break numbers for respective
explanatory variables}

\item{methoddisc}{A character vector of discretization methods}
}
\value{
A list of process and results of optimal interaction for
GD model.
}
\description{
Selecting optimal interaction for GD model.
}
\examples{
s1 <- selectgd(formula = y ~ xa + xb + xc, data = sim,
               ndisc = c(4,6,6), methoddisc = "quantile")

}
