% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim}
\alias{sim}
\title{Simulation data.}
\format{
\code{sim}: A data frame with 713 rows and 7 variables
}
\usage{
sim
}
\description{
Simulation data.
}
\author{
Yongze Song \email{yongze.song@postgrad.curtin.edu.au}
}
\keyword{dataset}
\keyword{sim}
